Wire House of Darkness into the minigame registry
