--- src/minigames/registryWrapperDraft.ts
@@ -1,1 +1,0 @@
-export * from './registryHouseOfDarkness';
